[UPDATE] Character 의 ATK를 float 으로 바꿈, Player의 레벨업 메서드 추가 (-성림)

+++ WeGoMars/WeGoMars/Character.cs
@@ -14,7 +14,7 @@ namespace WeGoMars{
 		public string Name { get; set; }
 		public string Job { get; set; }
 		public int Level { get; set; }
-		public int Atk { get; set; }
+		public float Atk { get; set; }
 		public int Def { get; set; }
 		public int MaxHp { get; set; }
 		public int MaxMp { get; set; }
